Rewrite GTO initial guess
